Tests: Fix block icons end to end test
When we insert a block it gets focused, so now we focus the nested child block that is inserted in the test blocks. So we are sure the selection of the parent block happens after events that may change the focus.
Before if the tests executed at a slow speed the focus may have been the child instead of the parent required for this tests.

We make the regex used here non greedy.